Display memory to dashboard
In order to save memory to database, this commit make migration from models and register it to admin with a little config to show un-editable id field.

This commit also handle the logic showing the marker of the memory place to the map and load memory's name and comments when click on the marker.

Lastly, this commit re-style some elements on tepmlates for better UI.